Added check to verify the file exists before removing it

+++ tasks/upload-metrics-tile/task.sh
@@ -19,5 +19,9 @@ SC_FILE_PATH=`find ./ -name *.tgz`
 
 ./om-cli/om-linux -t https://$OPS_MGR_HOST -u $OPS_MGR_USR -p $OPS_MGR_PWD -k upload-stemcell -s $SC_FILE_PATH
 
-echo "Removing downloaded stemcell $STEMCELL_VERSION"
-rm $SC_FILE_PATH
+if [ ! -f $SC_FILE_PATH ]; then
+    echo "Stemcell file not found!"
+else
+  echo "Removing downloaded stemcell $STEMCELL_VERSION"
+  rm $SC_FILE_PATH
+fi
